requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1625395912402/work
blinker==1.4
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///D:/bld/cffi_1625835469330/work
chardet @ file:///D:/bld/chardet_1610093614045/work
cryptography @ file:///D:/bld/cryptography_1616851634866/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dj-database-url==0.5.0
Django==3.1.13
django-allauth==0.41.0
django-extensions @ file:///home/conda/feedstock_root/build_artifacts/django-extensions_1618863301971/work
django-heroku==0.3.1
django-widget-tweaks==1.4.5
gunicorn==20.1.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
mkl-fft==1.3.0
mkl-random==1.0.4
mkl-service==2.3.0
numpy==1.17.0
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1622563202229/work
psycopg2==2.9.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1619620644068/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
PySocks @ file:///D:/bld/pysocks_1610291695773/work
python3-openid @ file:///home/conda/feedstock_root/build_artifacts/python3-openid_1612988842750/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1595492159598/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1602142927465/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
whitenoise==5.3.0
win-inet-pton @ file:///D:/bld/win_inet_pton_1610147511403/work
wincertstore==0.2
